Do not cache DB::NOW to avoid stale data in daemons
This feature reared its ugly head as a bug wherein the front page would
not update. Normally, the front page is updated whenever a season-level
update is requested and the season in question is "the current one".
This evaluation is done by fetching the current season, which means
fetching the current date. Because DB::NOW is cached the very first
time, long-lived scripts like the Daemon would be using the DateTime
when they were launched, which in this case was several months in the
past, during a previous season.

The action item is to ONLY use DB::T(DB::NOW) when the only interest is
on the DateTime class, i.e. in DBObject::db_field() responses.
Elsewhere, using the real DateTime is not only less error-prone, but
much more correct.

req https://trello.com/c/reChAYN4/242-in-progess-on-home-page