DBM: Do not issue order clause for subqueries
This was causing a significant performance issue discovered late on
Saturday, Oct 29, 2016. Specifically, the Regatta::getScoredRaces()
method, which is of the form:

  select * from race where ... and id in (
    select race from finish ORDER BY ENTERED
  ) order by ...;

Removing the pre-ordering the subquery (i.e. BY ENTERED) improved
performance significantly. It also makes sense: we always knew that
pre-ordering the internal query was unnecessary: now we have data that
it is actually harmful.