Fix DateTime::__construct() fatal exception bug
This one is on PHP: the DateTime constructor will throw a **fatal**
exception if the string provided is invalid. This exception cannot be
caught, even with a try-catch around it. We use the hack suggested in
the comments on the PHP manual of testing for validity with
`strtotime` prior to passing to constructor.

Maybe PHP will wisen up in the future, but as of PHP7, this is an
issue. Unit tests have been added, too.