Single Writer; Conf::$WRITER_PARAMS; configurable root for LocalHtmlWriter
The notion of multiple writers is deprecated. If multiple writers is
desired, a single implementation that delegates to multiple writers
can be created instead. By specifying a single writer in `Conf`, it is
possible to also specify a single set of configuration parameters, as
done for ERROR_HANDLER, etc.

To illustrate this usage, LocalHtmlWriter is updated to have a
configurable HTML root directory, instead of relying on inheritance.